fix(arena): drop self-fetch loop in /api/arena/start
The admin "Open lobby" button hit a 5xx (read as 404) because
/api/arena/start did a server-side fetch to ${ctx.url.origin}/api/arena/token —
hairpin loop from the website pod back to https://web.mentolder.de fails
inside the cluster (curl returns status 000 from the pod).

Extract the Keycloak token-exchange into website/src/lib/arena-token.ts and
call it directly from both /api/arena/token (browser-facing) and
/api/arena/start (server-side). No more self-fetch; the session's
access_token is already in scope.
